Update statistics and handling operations

+++ src/modules/cleaner/operations/removeDirectory.ts
@@ -1,15 +1,25 @@
 import fs from 'fs/promises';
 import path from 'path';
 import { fileExists } from '../../../utils/fileExist.js';
+import { handleError } from '../../../utils/handleError.js';
+import { CleanerStatistics } from '../cleaner.const.js';
 
-export const removeDirectory = async (file: string) => {
+export const removeDirectory = async (file: string, statistics: CleanerStatistics) => {
   try {
-    await fileExists(file);
-  } catch (e) {
-    return null;
-  }
+    const exists = await fileExists(file).catch(() => false);
+    if (!exists) {
+      statistics.incrementStat('unchanged');
+      return null;
+    }
 
-  const fullPath = path.resolve(file);
-  await fs.rmdir(fullPath, { recursive: true });
-  return true;
+    const fullPath = path.resolve(file);
+    await fs.rmdir(fullPath, { recursive: true });
+
+    statistics.incrementStat('removed');
+    return true;
+  } catch (error) {
+    statistics.incrementStat('error');
+    handleError(error as Error);
+    return false;
+  }
 };
